refactor: 增加 Group 标签监控逻辑
